refactor: unify command printers and resolve structural residuals
- Introduce @kuralPatterns commandPrinter across 5 command display
  modules: rename renderScore/renderResults to printScore/printAdvise,
  extract printAudit wrapper, move printPlace into a new place/display.ts
- Move highlight.ts from shell/commands/brief/ to shell/ui/ as a
  reusable terminal primitive; rewrite JSDoc to drop brief-specific
  framing
- Drop the now-stale @kuralResidual on duplicates.ts (containments
  finding resolved upstream, no suppression param needed); add
  @kuralPatterns declarationLine on declarationStart/EndLine and
  @kuralHelper on audit/display.ts kindPrefix and nodeLabel
- Rewrite printBrief description to lead with the eight facet
  sections it composes, clearing its outlier residual
