+++ src/shell/commands/advise/display.ts
@@ -172,11 +172,13 @@ function printCuts(result: AdviseResult): void {
 }
 
 /**
- * Prints all engine results to stdout with step lists and threshold sections.
+ * Composes the advise command's merge-order steps and threshold-cut
+ * sections into the terminal advise readout for stdout.
  * @param results - Array of engine results to display
- * @kuralCauses writes formatted output to stdout
+ * @kuralCauses writes advise readout to stdout
+ * @kuralPatterns commandPrinter
  */
-function renderResults(results: AdviseResult[]): void {
+function printAdvise(results: AdviseResult[]): void {
   for (const result of results) {
     printDendrogram(result);
     printCuts(result);
@@ -217,4 +219,4 @@ function printAdviseFooter(): void {
   );
 }
 
-export { printAdviseFooter, renderResults };
+export { printAdvise, printAdviseFooter };
